Changed idea for channel API slightly
Now the API consists of three functions: 'channelWrite', 'channelRead', and
'condChannel' which is new. The 'condChannel' function behaives similarly to
'cond' in that it enables or disables an Atom's action depending on a
condtion. In this case the special condition is that the channel being
listened to has an unread message waiting.